display_setting: Remove a warning
Avoids:

warning C4267: 'initializing': conversion from 'size_t' to 'DWORD', possible loss of data

Value won't never be that huge to cause overflow (we are reading a string
in a static buffer, we don't expect an encyclopedia).

+++ vdagent/display_setting.cpp
@@ -290,7 +290,7 @@ bool DisplaySetting::disable_wallpaper()
 
 static bool RegReadString(HKEY key, const TCHAR *name, TCHAR *buffer, size_t buffer_len)
 {
-    DWORD value_size = buffer_len * sizeof(buffer[0]);
+    DWORD value_size = (DWORD) (buffer_len * sizeof(buffer[0]));
     DWORD value_type;
     LONG status;
